feature/EAO-119 - GAC: Cálculo del Monto del GAC según Tabla y Días de Mora

+++ fineract-provider/src/main/java/org/apache/fineract/custom/portfolio/gac/domain/Gac.java
@@ -24,6 +24,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.Version;
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -47,7 +48,7 @@ public class Gac extends AbstractAuditableWithUTCDateTimeCustom {
     private Integer maximumAgeDays;
 
     @Column(name = "percentage", nullable = true)
-    private Integer percentageValue;
+    private BigDecimal percentageValue;
 
     @Getter
     @ManyToOne
